+++ Framework/PWLocation.framework/Versions/A/Headers/PWLocation.h
@@ -1,4 +1,4 @@
- //
+//
 //  PWLocation.h
 //  PWLocation
 //
@@ -13,9 +13,10 @@
 #import <PWLocation/PWMSELocationManager.h>
 #import <PWLocation/PWSLLocationManager.h>
 #import <PWLocation/PWMockLocationManager.h>
+#import <PWLocation/PWFusedLocationManager.h>
 
 /**
- PWLocation is a location based SDK.
+ PWLocation is a location provider SDK. This SDK encompasses a variety of indoor location providers and can be used in conjunction with PWMapKit to delivery and display indoor location location updates.
  */
 
 @interface PWLocation : NSObject