+++ Framework/PWLocation.framework/Versions/A/Headers/PWLocationProtocol.h
@@ -10,12 +10,12 @@
 static const NSInteger kPWUnknownFloorID = NSNotFound;
 
 /**
- The `PWLocation` is an abstract protocol that defines an indoor location. This protocol is primarily used for directions and for displaying the user's location on an indoor map.
+ The `PWLocation` is an abstract protocol that defines an indoor location. This protocol is primarily used for user-requested directions and for displaying the user's location on an indoor map.
  */
 @protocol PWLocation <NSObject>
 
 /**
- The geographical coordinate information. (read-only)
+ The latitude and longitude coordinates of the indoor location. (read-only)
  */
 @property (readonly) CLLocationCoordinate2D coordinate;
 
@@ -27,7 +27,7 @@ static const NSInteger kPWUnknownFloorID = NSNotFound;
 
 /**
  The floor ID associated with the location. 
- @discussion If the location is outdoors, this value should be `kPWUnknownFloorID`. (read-only)
+ @discussion If the location is outdoors this value should be `kPWUnknownFloorID`. (read-only)
  */
 @property (readonly) NSInteger floorID;