+++ Framework/PWLocation.framework/Versions/A/Headers/PWMockLocationStep.h
@@ -10,9 +10,9 @@
 @class PWIndoorLocation;
 
 /**
- A `PWMockLocationStep` object represents a discreet mock location step, which includes the `PWIndoorLocation` and an `updateInterval` property. Typically, you won't initialize this class directly.
+ A `PWMockLocationStep` object represents a discrete mock location step, which includes the `PWIndoorLocation` and an `updateInterval` property. Typically, this class isn't initialize directly.
  
- This class is designed to be used as is and should not be subclassed.
+ This class is designed to be used as-is and should not be subclassed.
  */
 
 @interface PWMockLocationStep : NSObject
@@ -27,6 +27,12 @@
  */
 @property (readonly) NSTimeInterval updateInterval;
 
+
+/**
+ Initializes a `PWMockLocation` step object with the specified dictionary.
+ @param dictionary The dictionary which contains the `PWLocationStep` information.
+ @return Returns a `PWLocationStep` object.
+ */
 + (PWMockLocationStep *)unpack:(NSDictionary *)dictionary;
 
 @end